services/json-pdf-embedder/requirements.txt
flask
pymupdf
google-genai
google-cloud-aiplatform
python-dotenv
werkzeug
gunicorn
